STM32F4 based PS2 to MSX Keyboard Adapter
This is an adapter to attach a PS/2 keyboard to a MSX [or other computers that use a up to 15 (input) x 8 (reading) matrix] with 4 bit BCD input. This is based on STM32F401, with expandable capabilities; using LIBOPENCM3 library and as a bare metal design.
This unit has a highly configurable capability, as it has an user updateable database, without need to recompile and upload a new code, to translate from PC keyboard to MSX.Expandable capabilities means that USB is still available to future improvements.