Add files via upload
Database compiler, now it is not obligated to use with paid M$ Excel, as it is capable to run with Basic of Libre Office / Star Office Calc. You have to configure Calc to be compatible with VBA and set to run macros. It will insert "Option VBASupport 1" in "Módulo1". No modifications in code are needeed.
Just click in the keyboard picture of desired sheet,  choose Y_Dummy as F (0xF or 15: The colunm with no keys) if you are planning to use with MSX. (0xF or 15: The colunm with no keys in destination keyboard)