STM32F401Mini based PS/2 to MSX Keyboard Adapter
Now fully functional. This is a bare metal design adapter based on STM32F401Mini to attach a PS/2 keyboard to a MSX [or other computers that use a up to 15 (input) x 8 (reading) matrix] with 4 bit BCD input. It has expandable capabilities, using LIBOPENCM3 library.
This unit has a highly configurable capability, as it has an user updateable database, without need to recompile and upload a new code, to translate from PC keyboard to MSX. Expandable capabilities means that USB is still available to future improvements.