Database compiler now compatible with Libre Office
Optimized Database compiler, now you are not obligated to use the paid M$ Excel, as it is capable to run with Basic of Libre Office / Open Office / Star Office Calc.
You have to configure Calc to be compatible with VBA and set to run macros. It will insert "Option VBASupport 1" in "Módulo1". No modifications in code are necessary.
Just click in the keyboard picture of desired sheet,  choose Y_Dummy as F (0xF or 15: The colunm with no keys in destination keyboard) if you are planning to use with MSX and wait for a message box that ihd.hex has been saved.