Excel Database compiler Libre Office compatible
Optimized Database compiler, now you are not obligated to use the paid M$ Excel, as it is capable to run with Basic of Libre Office / Open Office / Star Office Calc.
You have to configure Calc to be compatible with VBA and set to run macros. It will insert "Option VBASupport 1" in "Módulo1". No modifications in code are necessary.
Just click in the keyboard picture of desired sheet,  choose Y_Dummy as F (0xF or 15: The colunm with no keys in destination keyboard) if you are planning to use with MSX and wait for a message box that ihd.hex has been saved.
Quick instructions to upload new database:
1) Create the Database file ihd.hex (Intel Hex Database), running macro by clicking on the picture of "Your MSX Definitions" or "Expert Definitions" sheet
2) Turn off and remove the PS/2 keyboard;
3) Plug an (3.3 to 5V) serial adapter configured as bit rate 115200, 8 data bits, parity n (no), 1 stop bit. ATENTION: DO NOT connect a V.24/RS-232 interface or its working voltage will burn your hardware!
4) Run your prefered terminal adapter software (in Linux, if you permit, I could suggest my lightweight modified version of TIO (https://github.com/evandrosouza-developer/tio), as it is capable to send ASCII files through serial;
5) Power on the PS/2 to MSX Keyboard Adapter (or turn on the MSX host if the adapter is already installed) and follow the instructions.